requirements.txt
git+https://github.com/shahbuland/PerceptualSimilarity
torchmetrics
einops
ema-pytorch
vector-quantize-pytorch
ray
omegaconf
wandb
kornia
datasets
huggingface-hub
python-dotenv